Add password reset functionality and improve Telegram bot handling
- Implement forgot password and reset password flow with email
- Add password_reset_tokens migration
- Improve Telegram bot service with graceful degradation
- Allow users to rejoin goals after dropping
- Enhance dashboard UI with primary buttons and emojis
- Fix team invitation endpoint and status loading

+++ app/Services/TeamInvitationService.php
@@ -90,7 +90,7 @@ class TeamInvitationService
         DB::beginTransaction();
 
         try {
-            $expiryDays = config('teams.invitation_expiry_days', 7);
+            $expiryDays = (int) config('teams.invitation_expiry_days', 7);
 
             $invitation = TeamInvitation::create([
                 'team_id' => $team->id,
